requirement.txt
pyyaml
passlib
bcrypt